Reviewed-on: https://git.cubetiqs.com/sonphumrin/attendance-mobile/pulls/24

+++ lib/core/model/summary_attendance_model.dart
@@ -2,7 +2,7 @@ import 'package:attendance_app/core/repositories/base_model.dart';
 
 class SummaryAttendanceModel extends BaseModel<SummaryAttendanceModel> {
   final String? id;
-  final double? date;
+  final int? date;
   final int? totalAttendance;
   final int? totalAbsent;
   final int? totalLeave;
